% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fetching.R
\name{fetch_api_data}
\alias{fetch_api_data}
\title{Fetch Data from Remote API}
\usage{
fetch_api_data()
}
\value{
None
}
\description{
Fetches vehicle and stops data from a remote API and stores it in the global environment.
}
\details{
This function fetches real-time vehicle data and stops data from a remote API and stores them in the global environment. It uses the 'httr' and 'jsonlite' packages to make HTTP requests and parse JSON responses.

The fetched vehicle data includes information such as line ID, line text, departure information, and geographical coordinates. The fetched stops data includes properties and coordinates of stops.

The fetched data is stored in the global environment with variable names 'vehicle_data' and 'stops_data_flat'.
}
\seealso{
\code{\link{plot_stops_api}}
\code{\link{plot_bus_locations_api}}
\code{\link{find_busiest_stops_api}}
\code{\link{get_available_buses}}
}
