% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{find_busiest_stops_api}
\alias{find_busiest_stops_api}
\title{Finds most common stops based on API data}
\usage{
find_busiest_stops_api(stop_data)
}
\arguments{
\item{stop_data}{A data frame with stop information}
}
\value{
A data frame with the most common stops
}
\description{
Finds most common stops based on API data
}
\details{
Finds the most common and busiest stops based on API data.
}
\seealso{
\code{\link{plot_stops_api}}
\code{\link{plot_stops}}
}
