% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{find_nearest_stop}
\alias{find_nearest_stop}
\title{Finds the nearest bus stop to a given point in GTFS data.}
\usage{
find_nearest_stop(lat, lon)
}
\arguments{
\item{lat}{Latitude of the point.}

\item{lon}{Longitude of the point.}
}
\value{
Data for the nearest bus stop.
}
\description{
Finds the nearest bus stop to a given point in GTFS data.
}
\details{
Finds the nearest bus stop to a given point from GTFS data.
Given a Latitude and longitude, returns the information about the nearest bus stop.
}
\seealso{
\code{\link{plot_stops_api}}
\code{\link{plot_stops}}
}
