% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_available_buses}
\alias{get_available_buses}
\title{Find available bus lines for a given bus stop from API data.}
\usage{
get_available_buses("stop_name")
}
\arguments{
\item{stop_name}{The name of the bus stop}

\item{stops_data_flat}{A data frame containing flat stop location information}
}
\value{
Data for the available bus lines from the bus stop.
}
\description{
Find available bus lines for a given bus stop from API data.
}
\details{
Find available bus lines for a given bus stop from API
and returns the information about the buses at that stop and time.
}
\seealso{
\code{\link{plot_stops_api}}
\code{\link{plot_bus_locations_api}}
}
