% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_bus_locations_api}
\alias{plot_bus_locations_api}
\title{Creates a Leaflet map showing vehicle locations from API.}
\usage{
plot_bus_locations_api()
}
\value{
A Leaflet map object
}
\description{
Creates a Leaflet map showing vehicle locations from API.
}
\details{
Creates a Leaflet map showing bus locations from the given API data.
display live location of the vehicle
}
\seealso{
\code{\link{get_available_buses}}
}
