% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_specific_route}
\alias{plot_specific_route}
\title{Creates a Leaflet map displaying a specific route from GTFS data.}
\usage{
plot_specific_route(route_id)
}
\arguments{
\item{route_id}{The ID of the route to display}
}
\value{
A Leaflet map object
}
\description{
Creates a Leaflet map displaying a specific route from GTFS data.
}
\details{
Creates a Leaflet map displaying a specific route when given a bus number(route_id)
displays from GTFS data
}
\seealso{
\code{\link{plot_bus_locations_api}}
}
