% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_stops}
\alias{plot_stops}
\title{Creates a Leaflet map showing stop locations from GTFS data.}
\usage{
plot_stops(stops_data)
}
\arguments{
\item{stops_data}{A data frame containing stop locations}
}
\value{
A Leaflet map object
}
\description{
Creates a Leaflet map showing stop locations from GTFS data.
}
\details{
Creates a Leaflet map showing bus stop locations from the given GTFS data
}
\seealso{
\code{\link{plot_stops_api}}
}
