% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_stops_api}
\alias{plot_stops_api}
\title{Creates a Leaflet map showing stop locations from API.}
\usage{
plot_stops_api()
}
\arguments{
\item{stops_data_flat}{A data frame containing flat stop location information}
}
\value{
A Leaflet map object
}
\description{
Creates a Leaflet map showing stop locations from API.
}
\details{
Creates a Leaflet map showing bus stop locations from the given API data
}
\seealso{
\code{\link{plot_stops}}
}
