% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fetching.R
\name{read_gtfs_files}
\alias{read_gtfs_files}
\title{Reads GTFS text files and prepares data for analysis.}
\usage{
read_gtfs_files(gtfs_path)
}
\arguments{
\item{gtfs_path}{Path to the directory containing GTFS text files}
}
\value{
A list of data tables containing GTFS data and
}
\description{
Reads GTFS text files and prepares data for analysis.
}
